moved `buildFilename` function into helpers.js
