+++ utils/Loader.js
@@ -149,35 +149,6 @@ class Loader {
     }
 
     /**
-     * @desc A function which takes a component, which is an Object which
-     *       contains only one (1) property. The function will `switch` over
-     *       the property of the component and act accordingly. The function
-     *       will return a string representing the filename for the
-     *       configuration file for that component.
-     * @summary Construct the full filename of a given component.
-     * @param  {Object} component           Object with component information
-     * @param  {String} [component.service] The name portion of the service to
-     *                                      build a filename for.
-     * @return {String}                     Filename for the config of the
-     *                                      given component.
-     */
-    function buildFilename (component) {
-      if (Object.keys(component).length > 1) {
-        throw new CollinsError('Invalid:Input', {
-          details: 'too many properties on input object'
-        });
-      }
-      switch (Object.keys(component).join().toLowerCase()) {
-        case 'service':
-          return component.service + '.config.js';
-        default:
-          throw new CollinsError('Invalid:Input', {
-            details: 'invalid function params'
-          });
-      }
-    }
-
-    /**
      * @desc Sorts an array of configuration files for the Collins system.
      *       Each filename follows a paradigm. The first being
      *       `serviceGearName.config.js` for the name of a configuration file
